Update db debugging traffics

+++ Server/src/main/java/com/oose2016/group4/server/Crime.java
@@ -45,7 +45,7 @@ public class Crime {
 		return type;
 	}
 	
-	public double getLinkId() {
+	public int getLinkId() {
 		return linkid;
 	}
